requirements.txt
requests
BeautifulSoup4
wordcloud
numpy
pillow
